changed it again
Because first it did not really worked, not exactly sure why. So I had to change some things in both the transmit and the receive side of the code, the reason why it changed is also because I changed the functions we are using here, for example I set the standard channel that is used to 100 instead of forcing the user to do it.