changed some SPI commands
Because the code is currently not working for transmitting and receiving while it should work according to datasheets.
I decided to look into my choises for the SPI commands. (I did not yet tested if it works now because I am not done yet)
Apperantly I think I did indeed something wrong with the spi commands in write_payload, because after I send the W_TX_PAYLOAD_NOACK I setted the csn pin high, while I was not done with transmitting, because I had to send the value I want to send also in it.