checkRXfifo
to optimalize the code I wrote a function that checks if there is something on the receivers FIFO, so you won't be reading empty 0's that do not exist. So to use it make a if statement and just put the function in it and it wil work.
I also deleted a little mistake in the read function.